Changed terminology on line 40
Line 40 talks of pages and Action Panes, but these aren't Dataverse terminology or correct terminology in this context. Action Pages are AX based "ribbons" where as Dataverse has a "command bar":
https://learn.microsoft.com/en-us/power-apps/maker/model-driven-apps/command-designer-overview

I've changed Action Pane to command bar and have removed the reference to a "page", which is incorrect terminology as opening a project lands us on a "form". I've reworded the sentence to describe that the command bar should be selected when on a project's main form.